+++ src/api/l_physics_shapes.c
@@ -575,12 +575,9 @@ static int l_lovrConvexShapeGetScale(lua_State* L) {
 
 static int l_lovrConvexShapeSetScale(lua_State* L){
   ConvexShape* convex = luax_checktype(L,1,ConvexShape);
-
   float scale[3];
-  luax_readscale(L,2,scale,3,NULL);
-
-  lovrConvexShapeSetScale(convex,scale);
-
+  luax_readscale(L, 2, scale, 3, NULL);
+  luax_assert(L, lovrConvexShapeSetScale(convex, scale));
   return 0;
 }
 
@@ -607,12 +604,9 @@ static int l_lovrMeshShapeGetScale(lua_State* L) {
 
 static int l_lovrMeshShapeSetScale(lua_State* L) {
   MeshShape* mesh = luax_checktype(L, 1, MeshShape);
-
   float scale[3];
-  luax_readscale(L,2,scale,3,NULL);
-
-  lovrMeshShapeSetScale(mesh,scale);
-
+  luax_readscale(L, 2, scale, 3, NULL);
+  luax_assert(L, lovrMeshShapeSetScale(mesh, scale));
   return 0;
 }