+++ src/modules/physics/physics.c
@@ -2376,10 +2376,10 @@ void lovrConvexShapeGetScale(ConvexShape* shape, float* scale) {
   vec3_fromJolt(scale, &v);
 }
 
-void lovrConvexShapeSetScale(ConvexShape* convex, float* scale){
+bool lovrConvexShapeSetScale(ConvexShape* convex, float* scale){
   const JPH_Shape* hull = JPH_DecoratedShape_GetInnerShape((const JPH_DecoratedShape*) convex->handle);
   JPH_Shape* scaledHull = (JPH_Shape*) JPH_ScaledShape_Create(hull, vec3_toJolt(scale));
-  lovrShapeReplace(convex, scaledHull);
+  return lovrShapeReplace(convex, scaledHull);
 }
 
 MeshShape* lovrMeshShapeCreate(uint32_t vertexCount, float* vertices, uint32_t indexCount, uint32_t* indices, float* scale) {
@@ -2432,10 +2432,10 @@ void lovrMeshShapeGetScale(MeshShape* shape, float* scale) {
   vec3_fromJolt(scale, &v);
 }
 
-void lovrMeshShapeSetScale(MeshShape* shape, float* scale){
+bool lovrMeshShapeSetScale(MeshShape* shape, float* scale){
   const JPH_Shape* mesh = JPH_DecoratedShape_GetInnerShape((const JPH_DecoratedShape*) shape->handle);
   JPH_Shape* scaledMesh = (JPH_Shape*) JPH_ScaledShape_Create(mesh, vec3_toJolt(scale));
-  lovrShapeReplace(shape, scaledMesh);
+  return lovrShapeReplace(shape, scaledMesh);
 }
 
 TerrainShape* lovrTerrainShapeCreate(float* vertices, uint32_t n, float scaleXZ, float scaleY) {