+++ CHANGES.md
@@ -8,6 +8,7 @@ dev
 
 - Add `lovr.timer.getDisplayTime`.
 - Add support for atomic floats in shaders.
+- Add `ConvexShape:setScale` and `MeshShape:setScale`.
 
 ### Change